+++ apps/docs/src/theme/Playground/sandbox/templateFiles.ts
@@ -33,6 +33,8 @@ export const PACKAGE_JSON = JSON.stringify(
       '@coinbase/cds-common': 'latest',
       '@coinbase/cds-icons': 'latest',
       '@coinbase/cds-illustrations': 'latest',
+      '@coinbase/cds-lottie-files': 'latest',
+      '@coinbase/cds-utils': 'latest',
       'framer-motion': '^10.18.0',
     },
     devDependencies: {