+++ packages/web/src/dropdown/DropdownContent.tsx
@@ -1,4 +1,4 @@
-import React, { forwardRef, memo } from 'react';
+import { forwardRef, memo } from 'react';
 import {
   animateDropdownOpacityInConfig,
   animateDropdownOpacityOutConfig,
@@ -16,6 +16,10 @@ import type { DropdownProps } from './DropdownProps';
 
 const dropdownStaticClassName = 'cds-dropdown';
 
+/**
+ * @deprecated Use PopoverPanelContent within a PopoverPanel for interactive overlay content, or Select / SelectChip when presenting a list of selectable options. This will be removed in a future major release.
+ * @deprecationExpectedRemoval v10
+ */
 export type DropdownContentProps = {
   height?: React.CSSProperties['height'];
   placement?: Placement;
@@ -23,6 +27,10 @@ export type DropdownContentProps = {
 
 const MotionVStack = motion(VStack);
 
+/**
+ * @deprecated Use PopoverPanelContent within a PopoverPanel for interactive overlay content, or Select / SelectChip when presenting a list of selectable options. This will be removed in a future major release.
+ * @deprecationExpectedRemoval v10
+ */
 export const DropdownContent = memo(
   forwardRef<HTMLDivElement, DropdownContentProps>(
     ({ children, placement, minWidth = 'min-content', ...props }, ref) => {