+++ packages/common/CHANGELOG.md
@@ -8,6 +8,18 @@ All notable changes to this project will be documented in this file.
 
 <!-- template-start -->
 
+## 8.66.0 ((4/16/2026, 01:57 PM PST))
+
+This is an artificial version bump with no new change.
+
+## 8.65.0 ((4/16/2026, 10:06 AM PST))
+
+This is an artificial version bump with no new change.
+
+## 8.64.5 ((4/16/2026, 06:50 AM PST))
+
+This is an artificial version bump with no new change.
+
 ## 8.64.4 ((4/10/2026, 01:20 PM PST))
 
 This is an artificial version bump with no new change.